+++ src/pythonwrench/time.py
@@ -1,0 +1,31 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+import time
+from typing import Callable, Optional
+
+
+class Ticker:
+    def __init__(
+        self,
+        *,
+        get_time_fn: Callable[[], float] = time.perf_counter,
+        prev_tick: Optional[float] = None,
+    ) -> None:
+        if prev_tick is None:
+            prev_tick = get_time_fn()
+
+        super().__init__()
+        self._get_time_fn = get_time_fn
+        self._prev_tick = prev_tick
+
+    def tick(self) -> float:
+        now = self._get_time_fn()
+        duration = now - self._prev_tick
+        self._prev_tick = now
+        return duration
+
+    def set_prev_tick(self, prev_tick: Optional[float] = None) -> None:
+        if prev_tick is None:
+            prev_tick = self._get_time_fn()
+        self._prev_tick = prev_tick